+++ packages/ice/src/utils/routeManifest.ts
@@ -1,0 +1,26 @@
+import type { NestedRouteManifest } from '@ice/route-manifest';
+import getRoutePath from './getRoutePaths.js';
+
+export default class RouteManifest {
+  private routeManifest: NestedRouteManifest[];
+
+  constructor() {
+    this.routeManifest = null;
+  }
+
+  getNestedRoute() {
+    if (this.routeManifest === null) {
+      throw new Error('routeManifest is not initialized, call API in hooks.');
+    }
+    return this.routeManifest;
+  }
+
+  setRoutes(routes: NestedRouteManifest[]) {
+    console.log('routes ==>', routes);
+    this.routeManifest = routes;
+  }
+
+  getFlattenRoute() {
+    return getRoutePath(this.getNestedRoute());
+  }
+}