+++ packages/ice/CHANGELOG.md
@@ -3,7 +3,8 @@
 ## v3.1.1
 
 - [fix] value priority of host, port and https
-- [fix] The imported dataloaderFetcher name is error
+- [fix] the lanUrlForConfig should set when set host
+- [fix] the imported dataloaderFetcher name is error
 - [fix] the value of `isServer` is false when compile data-loader
 
 ## v3.1.0