+++ packages/ice/CHANGELOG.md
@@ -1,18 +1,27 @@
 # Changelog
 
+## v3.1.1
+
+- [feat] redirect runtime imports for data loader
+- [fix] the lanUrlForConfig should set when set host
+- [fix] the imported dataloaderFetcher name is error
+- [fix] the value of `isServer` is false when compile data-loader
+
 ## v3.1.0
 
+- [feat] export suspense api
+- [feat] refactor ability of keep target code
+- [feat] support render js bundle as entry 
+- [fix] support cli options of target, platform not is deprecated
 - [fix] build data loader with correct browser list
 - [fix] process exit when build with errors on dev
-- [feat] redirect runtime imports for data loader
-- [feat] optimize log info
+- [fix] optimize log info
+- [fix] should not replace global env for ssr
 
 ## v3.0.6
 
-- [feat] export suspense api
 - [feat] support `import.meta.target`, `import.meta.renderer` and `import.meta.env`
 - [feat] support config `routes.injectInitialEntry` for memory router
-- [fix] should not replace global env for ssr
 - [fix] cli option `--platform` change to `--target`
 - [fix] remove code when dead code is `ObjectProperties`
 - [fix] rebuild server entry when document changed